+++ src/services/sugestionServices/useOutsideClick.ts
@@ -1,0 +1,20 @@
+import { useEffect, useRef } from "react";
+
+export const useOutsideClick = (callback: () => void) => {
+    const ref = useRef<HTMLUListElement | null>(null);
+
+    useEffect(() => {
+        const handleClickOutside = (event: MouseEvent) => {
+            if (ref.current && !ref.current.contains(event.target as Node)) {
+                callback();
+            }
+        };
+
+        document.addEventListener("mousedown", handleClickOutside);
+        return () => {
+            document.removeEventListener("mousedown", handleClickOutside);
+        };
+    }, [callback]);
+
+    return ref;
+};
